+++ stdlib/LinearAlgebra/src/LinearAlgebra.jl
@@ -131,7 +131,7 @@ export
     rank,
     rdiv!,
     schur,
-    schurfact!,
+    schur!,
     svd,
     svdfact!,
     svdfact,