+++ NEWS.md
@@ -703,7 +703,8 @@ Deprecated or removed
   * `lufact`, `schurfact`, `lqfact`, `qrfact`, and `ldltfact` have respectively
     been deprecated to `lu`, `schur`, `lq`, `qr`, and `ldlt` ([#27159]).
 
-  *  `lufact!` has been deprecated to `lu!` ([#27159]).
+  * `lufact!` and `schurfact!` have respectively been deprecated to
+    `lu!` and `schur!` ([#27159]).
 
   * Indexing into multidimensional arrays with more than one index but fewer indices than there are
     dimensions is no longer permitted when those trailing dimensions have lengths greater than 1.